+++ packages/react-router/lib/dom/ssr/single-fetch.tsx
@@ -1,12 +1,11 @@
 import * as React from "react";
 import { decode } from "turbo-stream";
 import type { Router as DataRouter } from "../../router/router";
-import { isResponse, runMiddlewarePipeline } from "../../router/router";
+import { isResponse } from "../../router/router";
 import type {
   DataStrategyFunction,
   DataStrategyFunctionArgs,
   DataStrategyResult,
-  DataStrategyMatch,
 } from "../../router/utils";
 import {
   ErrorResponseImpl,
@@ -18,10 +17,10 @@ import {
 import { createRequestInit } from "./data";
 import type { AssetsManifest, EntryContext } from "./entry";
 import { escapeHtml } from "./markup";
-import type { RouteModule, RouteModules } from "./routeModules";
 import invariant from "./invariant";
-import type { EntryRoute } from "./routes";
 import { SINGLE_FETCH_REDIRECT_STATUS } from "../../server-runtime/single-fetch";
+import type { RouteModules } from "./routeModules";
+import type { DataRouteMatch } from "../../context";
 
 export const SingleFetchRedirectSymbol = Symbol("SingleFetchRedirect");
 
@@ -33,15 +32,22 @@ export type SingleFetchRedirectResult = {
   replace: boolean;
 };
 
+// Shared/serializable type used by both turbo-stream and RSC implementations
+type DecodedSingleFetchResults =
+  | { routes: { [key: string]: SingleFetchResult } }
+  | { redirect: SingleFetchRedirectResult };
+
+// This and SingleFetchResults are only used over the wire, and are converted to
+// DecodedSingleFetchResults in `fetchAndDecode`.  This way turbo-stream/RSC
+// can use the same `unwrapSingleFetchResult` implementation.
 export type SingleFetchResult =
   | { data: unknown }
   | { error: unknown }
   | SingleFetchRedirectResult;
 
-export type SingleFetchResults = {
-  [key: string]: SingleFetchResult;
-  [SingleFetchRedirectSymbol]?: SingleFetchRedirectResult;
-};
+export type SingleFetchResults =
+  | { [key: string]: SingleFetchResult }
+  | { [SingleFetchRedirectSymbol]: SingleFetchRedirectResult };
 
 interface StreamTransferProps {
   context: EntryContext;
@@ -50,6 +56,16 @@ interface StreamTransferProps {
   textDecoder: TextDecoder;
   nonce?: string;
 }
+
+// Some status codes are not permitted to have bodies, so we want to just
+// treat those as "no data" instead of throwing an exception:
+//   https://datatracker.ietf.org/doc/html/rfc9110#name-informational-1xx
+//   https://datatracker.ietf.org/doc/html/rfc9110#name-204-no-content
+//   https://datatracker.ietf.org/doc/html/rfc9110#name-205-reset-content
+//
+// Note: 304 is not included here because the browser should fill those responses
+// with the cached body content.
+export const NO_BODY_STATUS_CODES = new Set([100, 101, 204, 205]);
 
 // StreamTransfer recursively renders down chunks of the `serverHandoffStream`
 // into the client-side `streamController`
@@ -134,32 +150,65 @@ export function StreamTransfer({
   }
 }
 
-function handleMiddlewareError(error: unknown, routeId: string) {
-  return { [routeId]: { type: "error", result: error } };
-}
+type GetRouteInfoFunction = (match: DataRouteMatch) => {
+  hasLoader: boolean;
+  hasClientLoader: boolean;
+  hasShouldRevalidate: boolean;
+};
 
-export function getSingleFetchDataStrategy(
+type FetchAndDecodeFunction = (
+  args: DataStrategyFunctionArgs,
+  basename: string | undefined,
+  targetRoutes?: string[]
+) => Promise<{ status: number; data: DecodedSingleFetchResults }>;
+
+export function getTurboStreamSingleFetchDataStrategy(
+  getRouter: () => DataRouter,
   manifest: AssetsManifest,
   routeModules: RouteModules,
   ssr: boolean,
-  basename: string | undefined,
-  getRouter: () => DataRouter
+  basename: string | undefined
+): DataStrategyFunction {
+  let dataStrategy = getSingleFetchDataStrategyImpl(
+    getRouter,
+    (match: DataRouteMatch) => {
+      let manifestRoute = manifest.routes[match.route.id];
+      invariant(manifestRoute, "Route not found in manifest");
+      let routeModule = routeModules[match.route.id];
+      return {
+        hasLoader: manifestRoute.hasLoader,
+        hasClientLoader: manifestRoute.hasClientLoader,
+        hasShouldRevalidate: Boolean(routeModule?.shouldRevalidate),
+      };
+    },
+    fetchAndDecodeViaTurboStream,
+    ssr,
+    basename
+  );
+  return async (args) => args.unstable_runClientMiddleware(dataStrategy);
+}
+
+export function getSingleFetchDataStrategyImpl(
+  getRouter: () => DataRouter,
+  getRouteInfo: GetRouteInfoFunction,
+  fetchAndDecode: FetchAndDecodeFunction,
+  ssr: boolean,
+  basename: string | undefined
 ): DataStrategyFunction {
   return async (args) => {
     let { request, matches, fetcherKey } = args;
+    let router = getRouter();
 
     // Actions are simple and behave the same for navigations and fetchers
     if (request.method !== "GET") {
-      return runMiddlewarePipeline(
-        args,
-        false,
-        () => singleFetchActionStrategy(request, matches, basename),
-        handleMiddlewareError
-      ) as Promise<Record<string, DataStrategyResult>>;
+      return singleFetchActionStrategy(args, fetchAndDecode, basename);
     }
 
-    // TODO: Enable middleware for this flow
-    if (!ssr) {
+    let foundRevalidatingServerLoader = matches.some((m) => {
+      let { hasLoader, hasClientLoader } = getRouteInfo(m);
+      return m.unstable_shouldCallHandler() && hasLoader && !hasClientLoader;
+    });
+    if (!ssr && !foundRevalidatingServerLoader) {
       // If this is SPA mode, there won't be any loaders below root and we'll
       // disable single fetch.  We have to keep the `dataStrategy` defined for
       // SPA mode because we may load a SPA fallback page but then navigate into
@@ -192,71 +241,43 @@ export function getSingleFetchDataStrategy(
       //   errored otherwise
       // - So it's safe to make the call knowing there will be a `.data` file on
       //   the other end
-      let foundRevalidatingServerLoader = matches.some(
-        (m) =>
-          m.shouldLoad &&
-          manifest.routes[m.route.id]?.hasLoader &&
-          !manifest.routes[m.route.id]?.hasClientLoader
-      );
-      if (!foundRevalidatingServerLoader) {
-        return runMiddlewarePipeline(
-          args,
-          false,
-          () => nonSsrStrategy(manifest, request, matches, basename),
-          handleMiddlewareError
-        ) as Promise<Record<string, DataStrategyResult>>;
-      }
+      return nonSsrStrategy(args, getRouteInfo, fetchAndDecode, basename);
     }
 
     // Fetcher loads are singular calls to one loader
     if (fetcherKey) {
-      return runMiddlewarePipeline(
-        args,
-        false,
-        () => singleFetchLoaderFetcherStrategy(request, matches, basename),
-        handleMiddlewareError
-      ) as Promise<Record<string, DataStrategyResult>>;
+      return singleFetchLoaderFetcherStrategy(args, fetchAndDecode, basename);
     }
 
     // Navigational loads are more complex...
-    return runMiddlewarePipeline(
+    return singleFetchLoaderNavigationStrategy(
       args,
-      false,
-      () =>
-        singleFetchLoaderNavigationStrategy(
-          manifest,
-          routeModules,
-          ssr,
-          getRouter(),
-          request,
-          matches,
-          basename
-        ),
-      handleMiddlewareError
-    ) as Promise<Record<string, DataStrategyResult>>;
+      router,
+      getRouteInfo,
+      fetchAndDecode,
+      ssr,
+      basename
+    );
   };
 }
 
 // Actions are simple since they're singular calls to the server for both
 // navigations and fetchers)
 async function singleFetchActionStrategy(
-  request: Request,
-  matches: DataStrategyFunctionArgs["matches"],
+  args: DataStrategyFunctionArgs,
+  fetchAndDecode: FetchAndDecodeFunction,
   basename: string | undefined
 ) {
-  let actionMatch = matches.find((m) => m.shouldLoad);
+  let actionMatch = args.matches.find((m) => m.unstable_shouldCallHandler());
   invariant(actionMatch, "No action match found");
   let actionStatus: number | undefined = undefined;
   let result = await actionMatch.resolve(async (handler) => {
     let result = await handler(async () => {
-      let url = singleFetchUrl(request.url, basename);
-      let init = await createRequestInit(request);
-      let { data, status } = await fetchAndDecode(url, init);
+      let { data, status } = await fetchAndDecode(args, basename, [
+        actionMatch!.route.id,
+      ]);
       actionStatus = status;
-      return unwrapSingleFetchResult(
-        data as SingleFetchResult,
-        actionMatch!.route.id
-      );
+      return unwrapSingleFetchResult(data, actionMatch!.route.id);
     });
     return result;
   });
@@ -277,24 +298,29 @@ async function singleFetchActionStrategy(
 
 // We want to opt-out of Single Fetch when we aren't in SSR mode
 async function nonSsrStrategy(
-  manifest: AssetsManifest,
-  request: Request,
-  matches: DataStrategyFunctionArgs["matches"],
+  args: DataStrategyFunctionArgs,
+  getRouteInfo: GetRouteInfoFunction,
+  fetchAndDecode: FetchAndDecodeFunction,
   basename: string | undefined
 ) {
-  let matchesToLoad = matches.filter((m) => m.shouldLoad);
-  let url = stripIndexParam(singleFetchUrl(request.url, basename));
-  let init = await createRequestInit(request);
+  let matchesToLoad = args.matches.filter((m) =>
+    m.unstable_shouldCallHandler()
+  );
   let results: Record<string, DataStrategyResult> = {};
   await Promise.all(
     matchesToLoad.map((m) =>
       m.resolve(async (handler) => {
         try {
+          let { hasClientLoader } = getRouteInfo(m);
           // Need to pass through a `singleFetch` override handler so
           // clientLoader's can still call server loaders through `.data`
           // requests
-          let result = manifest.routes[m.route.id]?.hasClientLoader
-            ? await fetchSingleLoader(handler, url, init, m.route.id)
+          let routeId = m.route.id;
+          let result = hasClientLoader
+            ? await handler(async () => {
+                let { data } = await fetchAndDecode(args, basename, [routeId]);
+                return unwrapSingleFetchResult(data, routeId);
+              })
             : await handler();
           results[m.route.id] = { type: "data", result };
         } catch (e) {
@@ -309,121 +335,91 @@ async function nonSsrStrategy(
 // Loaders are trickier since we only want to hit the server once, so we
 // create a singular promise for all server-loader routes to latch onto.
 async function singleFetchLoaderNavigationStrategy(
-  manifest: AssetsManifest,
-  routeModules: RouteModules,
-  ssr: boolean,
+  args: DataStrategyFunctionArgs,
   router: DataRouter,
-  request: Request,
-  matches: DataStrategyFunctionArgs["matches"],
+  getRouteInfo: GetRouteInfoFunction,
+  fetchAndDecode: FetchAndDecodeFunction,
+  ssr: boolean,
   basename: string | undefined
 ) {
-  // Track which routes need a server load - in case we need to tack on a
-  // `_routes` param
+  // Track which routes need a server load for use in a `_routes` param
   let routesParams = new Set<string>();
 
-  // We only add `_routes` when one or more routes opts out of a load via
-  // `shouldRevalidate` or `clientLoader`
+  // Only add `_routes` when at least 1 route opts out via `shouldRevalidate`/`clientLoader`
   let foundOptOutRoute = false;
 
-  // Deferreds for each route so we can be sure they've all loaded via
-  // `match.resolve()`, and a singular promise that can tell us all routes
-  // have been resolved
-  let routeDfds = matches.map(() => createDeferred<void>());
-  let routesLoadedPromise = Promise.all(routeDfds.map((d) => d.promise));
+  // Deferreds per-route so we can be sure they've all loaded via `match.resolve()`
+  let routeDfds = args.matches.map(() => createDeferred<void>());
 
-  // Deferred that we'll use for the call to the server that each match can
-  // await and parse out it's specific result
-  let singleFetchDfd = createDeferred<SingleFetchResults>();
-
-  // Base URL and RequestInit for calls to the server
-  let url = stripIndexParam(singleFetchUrl(request.url, basename));
-  let init = await createRequestInit(request);
+  // Deferred we'll use for the singleular call to the server
+  let singleFetchDfd = createDeferred<DecodedSingleFetchResults>();
 
   // We'll build up this results object as we loop through matches
   let results: Record<string, DataStrategyResult> = {};
 
   let resolvePromise = Promise.all(
-    matches.map(async (m, i) =>
+    args.matches.map(async (m, i) =>
       m.resolve(async (handler) => {
         routeDfds[i].resolve();
+        let routeId = m.route.id;
+        let { hasLoader, hasClientLoader, hasShouldRevalidate } =
+          getRouteInfo(m);
 
-        let manifestRoute = manifest.routes[m.route.id];
+        let defaultShouldRevalidate =
+          !m.unstable_shouldRevalidateArgs ||
+          m.unstable_shouldRevalidateArgs.actionStatus == null ||
+          m.unstable_shouldRevalidateArgs.actionStatus < 400;
+        let shouldCall = m.unstable_shouldCallHandler(defaultShouldRevalidate);
 
-        // Note: If this logic changes for routes that should not participate
-        // in Single Fetch, make sure you update getLowestLoadingIndex above
-        // as well
-        if (!m.shouldLoad) {
-          // If we're not yet initialized and this is the initial load, respect
-          // `shouldLoad` because we're only dealing with `clientLoader.hydrate`
-          // routes which will fall into the `clientLoader` section below.
-          if (!router.state.initialized) {
-            return;
-          }
-
-          // Otherwise, we opt out if we currently have data and a
-          // `shouldRevalidate` function.  This implies that the user opted out
-          // via `shouldRevalidate`
-          if (
-            m.route.id in router.state.loaderData &&
-            manifestRoute &&
-            m.route.shouldRevalidate
-          ) {
-            if (manifestRoute.hasLoader) {
-              // If we have a server loader, make sure we don't include it in the
-              // single fetch .data request
-              foundOptOutRoute = true;
-            }
-            return;
-          }
+        if (!shouldCall) {
+          // If this route opted out, don't include in the .data request
+          foundOptOutRoute ||=
+            m.unstable_shouldRevalidateArgs != null && // This is a revalidation,
+            hasLoader && // for a route with a server loader,
+            hasShouldRevalidate === true; // and a shouldRevalidate function
+          return;
         }
 
         // When a route has a client loader, it opts out of the singular call and
         // calls it's server loader via `serverLoader()` using a `?_routes` param
-        if (manifestRoute && manifestRoute.hasClientLoader) {
-          if (manifestRoute.hasLoader) {
+        if (hasClientLoader) {
+          if (hasLoader) {
             foundOptOutRoute = true;
           }
           try {
-            let result = await fetchSingleLoader(
-              handler,
-              url,
-              init,
-              m.route.id
-            );
-            results[m.route.id] = { type: "data", result };
+            let result = await handler(async () => {
+              let { data } = await fetchAndDecode(args, basename, [routeId]);
+              return unwrapSingleFetchResult(data, routeId);
+            });
+
+            results[routeId] = { type: "data", result };
           } catch (e) {
-            results[m.route.id] = { type: "error", result: e };
+            results[routeId] = { type: "error", result: e };
           }
           return;
         }
 
         // Load this route on the server if it has a loader
-        if (manifestRoute && manifestRoute.hasLoader) {
-          routesParams.add(m.route.id);
+        if (hasLoader) {
+          routesParams.add(routeId);
         }
 
         // Lump this match in with the others on a singular promise
         try {
           let result = await handler(async () => {
             let data = await singleFetchDfd.promise;
-            return unwrapSingleFetchResults(data, m.route.id);
+            return unwrapSingleFetchResult(data, routeId);
           });
-          results[m.route.id] = {
-            type: "data",
-            result,
-          };
+          results[routeId] = { type: "data", result };
         } catch (e) {
-          results[m.route.id] = {
-            type: "error",
-            result: e,
-          };
+          results[routeId] = { type: "error", result: e };
         }
       })
     )
   );
 
   // Wait for all routes to resolve above before we make the HTTP call
-  await routesLoadedPromise;
+  await Promise.all(routeDfds.map((d) => d.promise));
 
   // We can skip the server call:
   // - On initial hydration - only clientLoaders can pass through via `clientLoader.hydrate`
@@ -436,26 +432,19 @@ async function singleFetchLoaderNavigationStrategy(
     (!router.state.initialized || routesParams.size === 0) &&
     !window.__reactRouterHdrActive
   ) {
-    singleFetchDfd.resolve({});
+    singleFetchDfd.resolve({ routes: {} });
   } else {
+    // When routes have opted out, add a `_routes` param to filter server loaders
+    // Skipped in `ssr:false` because we expect to be loading static `.data` files
+    let targetRoutes =
+      ssr && foundOptOutRoute && routesParams.size > 0
+        ? [...routesParams.keys()]
+        : undefined;
     try {
-      // When one or more routes have opted out, we add a _routes param to
-      // limit the loaders to those that have a server loader and did not
-      // opt out
-      if (ssr && foundOptOutRoute && routesParams.size > 0) {
-        url.searchParams.set(
-          "_routes",
-          matches
-            .filter((m) => routesParams.has(m.route.id))
-            .map((m) => m.route.id)
-            .join(",")
-        );
-      }
-
-      let data = await fetchAndDecode(url, init);
-      singleFetchDfd.resolve(data.data as SingleFetchResults);
+      let data = await fetchAndDecode(args, basename, targetRoutes);
+      singleFetchDfd.resolve(data.data);
     } catch (e) {
-      singleFetchDfd.reject(e as Error);
+      singleFetchDfd.reject(e);
     }
   }
 
@@ -466,34 +455,20 @@ async function singleFetchLoaderNavigationStrategy(
 
 // Fetcher loader calls are much simpler than navigational loader calls
 async function singleFetchLoaderFetcherStrategy(
-  request: Request,
-  matches: DataStrategyFunctionArgs["matches"],
+  args: DataStrategyFunctionArgs,
+  fetchAndDecode: FetchAndDecodeFunction,
   basename: string | undefined
 ) {
-  let fetcherMatch = matches.find((m) => m.shouldLoad);
+  let fetcherMatch = args.matches.find((m) => m.unstable_shouldCallHandler());
   invariant(fetcherMatch, "No fetcher match found");
-  let result = await fetcherMatch.resolve(async (handler) => {
-    let url = stripIndexParam(singleFetchUrl(request.url, basename));
-    let init = await createRequestInit(request);
-    return fetchSingleLoader(handler, url, init, fetcherMatch!.route.id);
-  });
+  let routeId = fetcherMatch.route.id;
+  let result = await fetcherMatch.resolve(async (handler) =>
+    handler(async () => {
+      let { data } = await fetchAndDecode(args, basename, [routeId]);
+      return unwrapSingleFetchResult(data, routeId);
+    })
+  );
   return { [fetcherMatch.route.id]: result };
-}
-
-function fetchSingleLoader(
-  handler: Parameters<
-    NonNullable<Parameters<DataStrategyMatch["resolve"]>[0]>
-  >[0],
-  url: URL,
-  init: RequestInit,
-  routeId: string
-) {
-  return handler(async () => {
-    let singleLoaderUrl = new URL(url);
-    singleLoaderUrl.searchParams.set("_routes", routeId);
-    let { data } = await fetchAndDecode(singleLoaderUrl, init);
-    return unwrapSingleFetchResults(data as SingleFetchResults, routeId);
-  });
 }
 
 function stripIndexParam(url: URL) {
@@ -539,11 +514,21 @@ export function singleFetchUrl(
   return url;
 }
 
-async function fetchAndDecode(
-  url: URL,
-  init: RequestInit
-): Promise<{ status: number; data: unknown }> {
-  let res = await fetch(url, init);
+async function fetchAndDecodeViaTurboStream(
+  args: DataStrategyFunctionArgs,
+  basename: string | undefined,
+  targetRoutes?: string[]
+): Promise<{ status: number; data: DecodedSingleFetchResults }> {
+  let { request } = args;
+  let url = singleFetchUrl(request.url, basename);
+  if (request.method === "GET") {
+    url = stripIndexParam(url);
+    if (targetRoutes) {
+      url.searchParams.set("_routes", targetRoutes.join(","));
+    }
+  }
+
+  let res = await fetch(url, await createRequestInit(request));
 
   // If this 404'd without hitting the running server (most likely in a
   // pre-rendered app using a CDN), then bubble a standard 404 ErrorResponse
@@ -553,44 +538,56 @@ async function fetchAndDecode(
 
   // Handle non-RR redirects (i.e., from express middleware)
   if (res.status === 204 && res.headers.has("X-Remix-Redirect")) {
-    let data: SingleFetchRedirectResult = {
-      redirect: res.headers.get("X-Remix-Redirect")!,
-      status: Number(res.headers.get("X-Remix-Status") || "302"),
-      revalidate: res.headers.get("X-Remix-Revalidate") === "true",
-      reload: res.headers.get("X-Remix-Reload-Document") === "true",
-      replace: res.headers.get("X-Remix-Replace") === "true",
+    return {
+      status: SINGLE_FETCH_REDIRECT_STATUS,
+      data: {
+        redirect: {
+          redirect: res.headers.get("X-Remix-Redirect")!,
+          status: Number(res.headers.get("X-Remix-Status") || "302"),
+          revalidate: res.headers.get("X-Remix-Revalidate") === "true",
+          reload: res.headers.get("X-Remix-Reload-Document") === "true",
+          replace: res.headers.get("X-Remix-Replace") === "true",
+        },
+      },
     };
-    if (!init.method || init.method === "GET") {
-      return {
-        status: SINGLE_FETCH_REDIRECT_STATUS,
-        data: { [SingleFetchRedirectSymbol]: data },
-      };
-    } else {
-      return { status: SINGLE_FETCH_REDIRECT_STATUS, data };
-    }
   }
 
-  // some status codes are not permitted to have bodies, so we want to just
-  // treat those as "no data" instead of throwing an exception.
-  // 304 is not included here because the browser should fill those responses
-  // with the cached body content.
-  const NO_BODY_STATUS_CODES = new Set([100, 101, 204, 205]);
   if (NO_BODY_STATUS_CODES.has(res.status)) {
-    if (!init.method || init.method === "GET") {
-      // SingleFetchResults can just have no routeId keys which will result
-      // in no data for all routes
-      return { status: res.status, data: {} };
-    } else {
-      // SingleFetchResult is for a singular route and can specify no data
-      return { status: res.status, data: { data: undefined } };
+    let routes: { [key: string]: SingleFetchResult } = {};
+    // We get back just a single result for action requests - normalize that
+    // to a DecodedSingleFetchResults shape here
+    if (targetRoutes && request.method !== "GET") {
+      routes[targetRoutes[0]] = { data: undefined };
     }
+    return {
+      status: res.status,
+      data: { routes },
+    };
   }
 
   invariant(res.body, "No response body to decode");
 
   try {
     let decoded = await decodeViaTurboStream(res.body, window);
-    return { status: res.status, data: decoded.value };
+    let data: DecodedSingleFetchResults;
+    if (request.method === "GET") {
+      let typed = decoded.value as SingleFetchResults;
+      if (SingleFetchRedirectSymbol in typed) {
+        data = { redirect: typed[SingleFetchRedirectSymbol] };
+      } else {
+        data = { routes: typed };
+      }
+    } else {
+      let typed = decoded.value as SingleFetchResult;
+      let routeId = targetRoutes?.[0];
+      invariant(routeId, "No routeId found for single fetch call decoding");
+      if ("redirect" in typed) {
+        data = { redirect: typed };
+      } else {
+        data = { routes: { [routeId]: typed } };
+      }
+    }
+    return { status: res.status, data };
   } catch (e) {
     // Can't clone after consuming the body via turbo-stream so we can't
     // include the body here.  In an ideal world we'd look for a turbo-stream
@@ -657,45 +654,42 @@ export function decodeViaTurboStream(
   });
 }
 
-function unwrapSingleFetchResults(
-  results: SingleFetchResults,
+function unwrapSingleFetchResult(
+  result: DecodedSingleFetchResults,
   routeId: string
 ) {
-  let redirect = results[SingleFetchRedirectSymbol];
-  if (redirect) {
-    return unwrapSingleFetchResult(redirect, routeId);
+  if ("redirect" in result) {
+    let {
+      redirect: location,
+      revalidate,
+      reload,
+      replace,
+      status,
+    } = result.redirect;
+    throw redirect(location, {
+      status,
+      headers: {
+        // Three R's of redirecting (lol Veep)
+        ...(revalidate ? { "X-Remix-Revalidate": "yes" } : null),
+        ...(reload ? { "X-Remix-Reload-Document": "yes" } : null),
+        ...(replace ? { "X-Remix-Replace": "yes" } : null),
+      },
+    });
   }
 
-  return results[routeId] !== undefined
-    ? unwrapSingleFetchResult(results[routeId], routeId)
-    : null;
-}
-
-function unwrapSingleFetchResult(result: SingleFetchResult, routeId: string) {
-  if ("error" in result) {
-    throw result.error;
-  } else if ("redirect" in result) {
-    let headers: Record<string, string> = {};
-    if (result.revalidate) {
-      headers["X-Remix-Revalidate"] = "true";
-    }
-    if (result.reload) {
-      headers["X-Remix-Reload-Document"] = "true";
-    }
-    if (result.replace) {
-      headers["X-Remix-Replace"] = "true";
-    }
-    throw redirect(result.redirect, { status: result.status, headers });
-  } else if ("data" in result) {
-    return result.data;
+  let routeResult = result.routes[routeId];
+  if ("error" in routeResult) {
+    throw routeResult.error;
+  } else if ("data" in routeResult) {
+    return routeResult.data;
   } else {
     throw new Error(`No response found for routeId "${routeId}"`);
   }
 }
 
 function createDeferred<T = unknown>() {
-  let resolve: (val?: any) => Promise<void>;
-  let reject: (error?: Error) => Promise<void>;
+  let resolve: (val: T) => Promise<void>;
+  let reject: (error: unknown) => Promise<void>;
   let promise = new Promise<T>((res, rej) => {
     resolve = async (val: T) => {
       res(val);
@@ -703,7 +697,7 @@ function createDeferred<T = unknown>() {
         await promise;
       } catch (e) {}
     };
-    reject = async (error?: Error) => {
+    reject = async (error?: unknown) => {
       rej(error);
       try {
         await promise;