+++ packages/create-react-router/CHANGELOG.md
@@ -1,6 +1,16 @@
 # `create-react-router`
 
+## 7.5.1
+
+_No changes_
+
+## 7.5.0
+
+_No changes_
+
 ## 7.4.1
+
+_No changes_
 
 ## 7.4.0